requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
nltk>=3.6
PyPDF2>=3.0.0
PyMuPDF>=1.20.0
sentence-transformers>=2.2.0
transformers>=4.20.0
torch>=1.11.0
faiss-cpu>=1.7.0
